backend/requirements.txt
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
firebase-admin==5.2.0
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.46.0
grpcio-status==1.46.0
gunicorn==20.1.0
h5py==3.6.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
msgpack==1.0.3
npm==0.1.1
npx==0.1.1
numpy==1.22.3
oauthlib==3.2.0
opencv-python-headless==4.5.5.64
openpyxl==3.1.2
opt-einsum==3.3.0
optional-django==0.1.0
packaging==21.3
pandas==2.0.0
Pillow==9.1.0
platformdirs==3.1.1
proto-plus==1.20.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
typing_extensions==4.2.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.21.0
Werkzeug==2.1.2
whitenoise==5.3.0
wrapt==1.14.1
zipp==3.8.0
